Made routes modular to better fit the folder structure.
