+++ src/index.js
@@ -1,35 +1,37 @@
-import Express from 'express';
-import cors from 'cors';
-import debug from 'debug';
-import bodyParser from 'body-parser';
-import routes from './routes/index';
-import { PORT } from './config/constants';
+const express = require('express');
+const mongoose = require('mongoose');
 
-/**
-* @fileOverview - application entry point
-* @requires - express
-* @requires - body-parser
-* @requires - cors
-* @requires - ./routes
-* @exports - index.js
-* */
+const bodyParser = require('body-parser');
+const cookieParser = require('cookie-parser');
 
-// declare constants
-const app = new Express();
-const port = PORT || 3000;
+const {DB_URL, PORT} = require('../config/constants.js');
 
-// declare middleware
-app.use(bodyParser.urlencoded({
-  extended: false,
-}));
+import initRoutes from './routes/routes';
+
+const app = express();
+
+mongoose.connect(DB_URL)
+  .then(() => {
+    console.log('Successfully connected to MongoDB Atlas!');
+  })
+  .catch((error) => {
+    console.log('Unable to connect to MongoDB Atlas!');
+    console.error(error);
+	});
+	
+app.set('view engine', 'ejs');
+
+app.use((req, res, next) => {
+  //res.setHeader('Access-Control-Allow-Origin', '*'); //Don't think we need CORS here.
+  res.setHeader('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content, Accept, Content-Type, Authorization');
+  res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, PATCH, OPTIONS');
+  console.log('Response headers set');
+  next();
+});
+
 app.use(bodyParser.json());
+app.use(cookieParser()); //Parse the cookie data (User ID).
 
-// 3rd party middleware
-app.use(cors('*'));
+initRoutes(app);
 
-routes(app);
-
-// listen to app port
-app.listen(port, debug('app')(`App listening on port ${port}`));
-
-export default app;
+app.listen(PORT, ()=> console.log(`Server listening on port ${PORT}`));
