+++ src/middlewares/middlewares.js
@@ -1,0 +1,5 @@
+import renderLandingPage from './renderLandingPage';
+import trimUrl from './trim-url';
+import deleteUrl from './delete-url';
+
+export default { renderLandingPage, trimUrl, deleteUrl };
