+++ src/middlewares/trim-url.js
@@ -1,0 +1,1 @@
+import UrlShorten from '../models/UrlShorten.js';
