+++ src/models/UrlShorten.js
@@ -1,13 +1,13 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
 const { Schema } = mongoose;
 
 const urlShortenSchema = new Schema({
-    long_url: String,
-    clipped_url: String,
-    urlCode: String,
-    created_by: String,
-    createdAt: { type: Date, default: Date.now },
-    click_count: Number,
+	long_url: { type: String, required: true },
+	clipped_url: { type: String, required: true },
+	urlCode: { type: String, required: true },
+	created_by: { type: String, required: true },
+	createdAt: { type: Date, default: Date.now },
+	click_count: { type: Number, required: true }
 });
 
-module.exports = mongoose.model("UrlShorten", urlShortenSchema);
+export default mongoose.model("UrlShorten", urlShortenSchema);
