--- src/routes/index.txt
@@ -1,1 +1,0 @@
-routes files goes here
